거의완성

+++ feelm/lib/View/movie_detail_screen.dart
@@ -1,4 +1,9 @@
+import 'dart:developer';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
+
+import 'package:feelm/View/mainScreen.dart';
+import 'package:feelm/View/oneline_reviews.dart';
 import 'package:feelm/json/movie_json.dart';
 import 'package:feelm/main.dart';
 
@@ -63,20 +68,52 @@ class _MovieDetailScreenState extends State<MovieDetailScreen> {
         .collection('favorite');
 
     if (isFavorite) {
-      //즐겨찾기 해제
-      await favoriteCollection.doc(widget.movieName).delete();
+      // 다이얼로그 표시
+      final shouldRemove = await showDialog<bool>(
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            title: const Text(
+              "이미 작성한 다이어리가 있다면 삭제됩니다.\n즐겨찾기를 해제하시겠습니까?",
+              style: TextStyle(fontSize: 16), // 폰트 크기 조정
+              textAlign: TextAlign.center,
+            ),
+            actionsAlignment: MainAxisAlignment.spaceEvenly, // 버튼 동일
+            actions: [
+              TextButton(
+                onPressed: () => Navigator.of(context).pop(false), // 취소 선택
+                child: const Text("취소"),
+              ),
+              TextButton(
+                onPressed: () => Navigator.of(context).pop(true), // 확인 선택
+                child: const Text("확인"),
+              ),
+            ],
+          );
+        },
+      );
+
+      // 사용자가 '확인'을 눌렀을 경우만 삭제 실행
+      if (shouldRemove == true) {
+        await favoriteCollection.doc(widget.movieName).delete();
+        setState(() {
+          isFavorite = false;
+        });
+      }
     } else {
-      //즐겨찾기 추가
+      // 즐겨찾기 추가
       await favoriteCollection.doc(widget.movieName).set({
         'title': selectedMovie!.title,
         'poster': selectedMovie!.poster,
         'genre': selectedMovie!.genre,
-        'runtime': selectedMovie!.runtime
+        'runtime': selectedMovie!.runtime,
+        'diaryText': "",
+        'savedDate': "",
+      });
+      setState(() {
+        isFavorite = true;
       });
     }
-    setState(() {
-      isFavorite = !isFavorite;
-    });
   }
 
   @override
@@ -281,9 +318,49 @@ class _MovieDetailScreenState extends State<MovieDetailScreen> {
                         ),
                       ),
 
-                      const SizedBox(
-                        height: 10,
-                      ),
+                      // const SizedBox(
+                      //   height: 10,
+                      // ),
+                      // Reviews 섹션 추가
+                      OnelineReviews(movieName: widget.movieName),
+
+                      Center(
+                        child: TextButton(
+                          onPressed: () {
+                            if (isFavorite) {
+                              log("record my feelm");
+                              Navigator.push(
+                                  context,
+                                  MaterialPageRoute(
+                                      builder: (context) =>
+                                          const MainScreen()));
+                            }
+                          },
+                          style: TextButton.styleFrom(
+                            padding: const EdgeInsets.symmetric(
+                                horizontal: 120, vertical: 10),
+                            backgroundColor: isFavorite
+                                ? const Color(0xFF757575)
+                                : Colors.red,
+                            shape: RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(30),
+                            ),
+                            elevation: 5, //그림자 효과
+                            shadowColor: Colors.black.withOpacity(0.2), //그림자 색상
+                          ),
+                          child: FittedBox(
+                            fit: BoxFit.scaleDown, // 텍스트가 버튼 안에 맞도록 축소
+                            child: Text(
+                              isFavorite ? 'record my feelm' : '즐겨찾기를 추가해주세요',
+                              style: const TextStyle(
+                                color: Colors.white,
+                                fontSize: 14,
+                                fontWeight: FontWeight.w500,
+                              ),
+                            ),
+                          ),
+                        ),
+                      )
                     ],
                   ),
                 ),
